regex_tool/requirements.txt
django
django-environ
djangorestframework
environ
flask
flask_pymongo
gunicorn
glob2
pandas
pymongo
PyPDF2
QtPy
werkzeug
wasabi
waitress
